pyproject.toml
[project]
name = "algogen"
version = "0.1.0"
description = "Multi-agent algorithm package generator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cyaron>=0.7.0",
    "langchain>=1.2.13",
    "langchain-deepseek>=1.0.1",
    "langgraph>=1.1.3",
    "pydantic>=2.11.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
]

[project.scripts]
algogen = "algogen.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/algogen"]
